wip: fix the texture applying immediately
