+++ assets/shaders/myshader_2d.wgsl
@@ -1,17 +1,26 @@
-//! Showing how to use a texture, drag-n-drop for you own texture will be supported soon.
+/// ***************************** ///
+/// THIS IS THE DEFAULT 2D SHADER ///
+/// You can always get back to this with `python3 scripts/reset-2d.py` ///
+/// ***************************** ///
 
-#import bevy_pbr::mesh_vertex_output      MeshVertexOutput
+#import bevy_pbr::mesh_vertex_output MeshVertexOutput
+#import bevy_sprite::mesh2d_view_bindings globals 
+#import shadplay::shader_utils::common NEG_HALF_PI, shaderToyDefault, rotate2D
 
-@group(1) @binding(1) var texture: texture_2d<f32>;
-@group(1) @binding(2) var texture_sampler: sampler;
+#import bevy_render::view  View
+@group(0) @binding(0) var<uniform> view: View;
+
+const SPEED:f32 = 1.0;
 
 @fragment
 fn fragment(in: MeshVertexOutput) -> @location(0) vec4<f32> {
     // ensure our uv coords match shadertoy/the-lil-book-of-shaders
-    let texture_uvs = in.uv;
+    var uv = (in.uv * 2.0) - 1.0;
+    let resolution = view.viewport.zw;
+    let t = globals.time * SPEED;
+    uv.x *= resolution.x / resolution.y;
+    uv *= rotate2D(NEG_HALF_PI);
 
-    let tex: vec4f = textureSample(texture, texture_sampler, texture_uvs); 
-
-    return tex;
+    return vec4f(shaderToyDefault(t, uv), 1.0);
 }
